feat: Architect a Blockchain Supply Chain Solution
This commit includes all changes required to complete the project. The work that has been done is best described by explaining its main areas:

1) Step 1  Planning of the project with write-ups

This is the first step of the project, and is made up of two different parts:

1.1) Step 1.1   Project write-up - UML

The first requirement of the first step of this project is the planning of the solution. With this aim in view, a number of UML documents have been generated:

* Activity diagram
* Sequence diagram
* State diagram
* Classes diagram

2.2) Step 1.2	Project write-up - Libraries

This project has made use of:

* Node v10.15.3
* Truffle v4.1.15
* Solidity v0.4.25
* web3 1.0.0-beta.37
* truffle-hdwallet-provider 1.0.13
* lite-server 2.4.0
* Ganache CLI v6.4.3 (ganache-core: 2.5.5)
* Metamask Version 6.7.2

2) Step 2	Writing the smart contracts

This is the second step of the project, and is made up of four different parts:

2.1) Step 2.1	Defining and implementing interfaces

The first requirement of the second step of this project focuses in the inheritance structure of the smart contracts. This has been established like this:

* contracts FarmerRole, DistributorRole, RetailerRole, and ConsumerRole, import library Roles.
* contract SupplyChain is FarmerRole, DistributorRole, RetailerRole, ConsumerRole.
* contract SupplyChain is Ownable.

2.2) Step 2.2   Building out AccessControl Contracts

In this requirement, the AccessControl contracts have been built out (FarmerRole.sol, DistributorRole.sol, RetailerRole.sol, and ConsumerRole.sol).

2.3) Step 2.3   Building out Base Contract

In this requirement, the Base Contract (SupplyChain.sol) is built out. It is noteworthy that the ownership management in this contract is not used, and the one in the Core Contract (ownable.sol) is used instead.

2.4) Step 2.4   Building out Core Contract

This requirement has already been provided in the starter code that this project intends to build on. As already mentioned, the ownership management used is the one in this contract.

3) Step 3  Testing smart contract code coverage

The third step of the project builds out the tests for all functions in the sequence diagram. Note that before the tests are executed, the different actors are assigned their roles in the system. This is done only once before the entire group of tests, not once before each test.

The eleven tests covered are:

* Testing smart contract function harvestItem() that allows a farmer to harvest coffee.
* Testing smart contract function processItem() that allows a farmer to process coffee.
* Testing smart contract function packItem() that allows a farmer to pack coffee.
* Testing smart contract function sellItem() that allows a farmer to sell coffee.
* Testing smart contract function buyItem() that allows a distributor to buy coffee.
* Testing smart contract function shipItem() that allows a distributor to ship coffee.
* Testing smart contract function receiveItem() that allows a retailer to mark coffee received.
* Testing smart contract function purchaseItem() that allows a consumer to purchase coffee.
* Testing smart contract function fetchItemBufferOne() that allows anyone to fetch item details from blockchain.
* Testing smart contract function fetchItemBufferTwo() that allows anyone to fetch item details from blockchain.
* Testing smart contract function transferOwnership() that allows the owner of the contract to transferownership.

4) Step 4	Deploying smart contracts on a public test network (Rinkeby)

This is the fourth step of the project, where the smart contract has been deployed to the Rinkeby public test network.

After successful deployment, the contract address hash is: 0x0aa5234ea0992000cb1836d4063993c775b26c63. It can be accessed on Etherscan at https://rinkeby.etherscan.io/address/0x0aa5234ea0992000cb1836d4063993c775b26c63. The transaction hash for contract creation is: 0x1e527e482ec7f4da662263ceffd0bd6bbdd3a0ed2470ca5cf97c9c18143e0f25.

5) Step 5  Modifying client code to interact with smart contracts

In this last step of the project, a front-end is developed that makes it possible for the different actors of the system to manage product items from origin to end within the supply chain.